+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 ### Features
 ### Enhancements
 - Clean up / hide complex fields on agent configuration ([#796](https://github.com/opensearch-project/dashboards-flow-framework/pull/796))
+- Add agent summary ([#801](https://github.com/opensearch-project/dashboards-flow-framework/pull/801))
 ### Bug Fixes
 ### Infrastructure
 ### Documentation